refactor: bmi_calculator: Update folder structure
