+++ lib/components/customized_cards.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'constants.dart';
+import '../constants.dart';
 
 class GenderCards extends StatelessWidget {
   String gender;
