+++ lib/main.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
-import 'input_page.dart';
-import 'result_page.dart';
+import 'screens/input_page.dart';
+import 'screens/result_page.dart';
 
 void main() {
   runApp(const BMICalculator());
