+++ lib/screens/result_page.dart
@@ -1,7 +1,7 @@
-import 'package:bmi_calculator/reusable_card.dart';
+import 'package:bmi_calculator/components/reusable_card.dart';
 import 'package:flutter/material.dart';
-import 'constants.dart';
-import 'bottom_buton.dart';
+import '../constants.dart';
+import '../components/bottom_buton.dart';
 
 class ResultPage extends StatelessWidget {
   @override
